requirements.txt
streamlit
opencv-python
pillow
pytesseract
pdf2image
requests
pandas
